Remove parameter validation
While a good idea in principal most parameters cannot be validated by just
comparing against a static list.
Firstly, the number or names of the valid options might change over time.
Secondly, Parameters like --digest have a dynamic list of permissible
values (e.g. dependant on the installed version of OpenSSL).

Doing this properly would take a lot of time, so I argue that it's better to
follow a "it will break on its own" philosophy.