Convert all types to work with a PKI directory structure
The easy-rsa interface and file locations have historically been in constant
change. Currently, it looks like the project is steering towards having a single
PKI directory and keeping CAs, key pairs and configuration within it.

The packaging in distributions hasn't caught up everywhere (e.g. Debian still
ships a make-cadir helper) or OpenWrt for some reason has a global
/etc/easy-rsa/vars.

Supporting different versions and packaging in this set would be a nightmere,
thus this commit tries to do update the __easy_rsa_* types to create and work
with a simple PKI directory structure as "suggested" by upstream.

Yes, this is a breaking change and no efforts were made to allow for a
transition between the old and new style.